Scope handling rework
Move scoping responisibilities to the Stmts AST node and modify
SymbolTable to become a simple stack of Stmts.
SymbolTable is no longer created at the compiler level, but instantiated
as needed by the parser and the semantic pass.
Stmts maintains an ID -> Expr map for identifiers and an ID -> Type
map for types. Stmts are pushed and popped off of the symbol table
during both parsing and semantic analysis. Types are added during
parsing, and IDs are added during the semnantic pass.

IDs may be local variable access (ID -> VarExpr) added when visiting a
VarDeclaration, field access (ID -> this.fieldname) added when
visiting an UnresolvedClassDefinition, or formal argument access
(ID -> VarExpr) added before resolving a Method's statements.  These
IDs are looked up when visiting an UnresolvedIdentifier.

The smenatic pass now maintains a currentMethod_, used when visiting a
return statement to validate the return type as well as to know how
far to unwind the stack. This relieves the symbol table of knowing
which type of scope is current (class, method, enum, or stmts).

UnresolvedClassType now holds a ClassType instead of a Scope. During
parsing, a list of statements is built for the class body, consisting
of VarDeclarations, and MethodDecls.  At the end of a class
declaration, these statments are converted into Fields and Methods via
the ClassPopulator visitor. For now, MethodDecls are only used during
parsing and are not encountered during type replacement or semantic
analysis.

Enums, formerly using the symbol table for scoping, now use their
own currentEnumType_ global in the parser.

The parser was modified use its own SymbolTable to maintain the
current scope. New statements are added to the enclosing scope in
AppendStatement(). The Bison stack is still used in part for non-
block statements such as if, else, while, etc bodies. This should
be cleaned up in the future.

The compilers now create a root Stmts node instead of a Scope, and
pass that to all passes.

Since the symbol table is no longer persistent and doesn't own its
scopes, SymbolTable::Dump() and the associated command-line flag were
removed.